config/settings.py
# config/settings.py
import os
from dotenv import load_dotenv

load_dotenv()

SUPABASE_URL = os.getenv("SUPABASE_URL")
SUPABASE_KEY = os.getenv("SUPABASE_KEY")
GMAIL_CREDENTIALS = os.getenv("GMAIL_CREDENTIALS_PATH")
ANTHROPIC_API_KEY = os.getenv("ANTHROPIC_API_KEY")
VOYAGE_API_KEY= os.getenv("VOYAGE_API_KEY")